Refactored all the core logic of AutoPanMix into new base classes (see below) and replaced it with three new specialized mixers that each have unique parameters and scaling characteristics:
* MonoPanWithPersonalMixer (squash to mono, pan and creates personal mixes)
* MonoPanWithoutPersonalMixer (squash to mono, pan without personal mixes)
* StereoMixer (flat mix stereo and creates personal mixes)

Renamed SimpleMix to SimpleMixer, but it is largely unchanged

Created a new InputBusMixer that establishes the private input bussses
and creates a synth to deliver audio to them. This is a new base class
for other mixers.

Created a new OutputBusMixer that creates a flat audio mix from the
private input busses, and delivers it to all clients. This is a new
base class for other mixers, which extends InputBusMixer.

Created a new PersonalMixer that creates a personal output mix for
each client by mixing together the audio from the private input busses,
and delivers audio to them. This is a new base class for other mixers,
which extends InputBusMixer

Renamed BaseMix to BaseMixer and adjusted it to support new signal
processing chains: preChain and postChain