Change Self Volume label to Studio Loopback Volume
Also adding "Studio" to "Master Volume"

Trying to address two issues here. One is that these controls impact
the entire studio, versus only YOU. The other is that use of "Self"
makes it sound like this is something that one would expect to be
able to adjust uniquely for each participant.