Removing PanningLink from default preset
This is no longer required now that we have "Mix to mono"
and automatically copy left to right (via Jack connections)
when only a single channel is sent.

It doesn't require people who want to send stereo mixes to
change their soundscapes presets to make it sound good, and
makes it easier to mix and match across a group.

Other presets, and of course custom presets are always
availabe for anyone who wants auto panning. And no existing
studio configurations will be changed unless they do it
explicitly.